requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
azure-ai-documentintelligence==1.0.0b2
azure-ai-formrecognizer==3.3.0
azure-common==1.1.28
azure-core==1.30.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.62
botocore==1.34.62
browser-cookie3==0.19.1
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==4.47
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.4
decorator==5.1.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastavro==1.9.4
fonttools==4.51.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.42
gotrue==2.1.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
isodate==0.6.1
jeepney==0.8.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.1.7
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.6
google-generativeai
langsmith==0.0.87
linkedin_api==2.1.1
loguru==0.7.2
lxml==5.1.0
lz4==4.3.3
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.2
mdurl==0.1.2
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
openai==1.12.0
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
parsel==1.9.1
pillow==10.2.0
plotly==5.18.0
postgrest==0.15.0
protobuf==4.25.3
pyarrow==15.0.0
pycryptodomex==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.8.2
python-jobspy==1.1.52
python-pptx==0.6.23
pytz==2024.1
PyYAML==6.0.1
rank-bm25==0.2.2
realtime==1.0.2
referencing==0.33.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.0
rpds-py==0.18.0
s3transfer==0.10.0
scrapfly-sdk==0.8.16
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.27
storage3==0.7.0
streamlit==1.32.0
streamlit-card==1.0.2
streamlit-modal==0.1.2
StrEnum==0.4.15
supabase==2.3.5
supafunc==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tls-client==1.0.1
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
w3lib==2.1.2
watchdog==4.0.0
websockets==11.0.3
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.17.0